app/main.py
from commands import *
from os import getenv
import logging
import google.cloud.logging

if (env := getenv("ENV")) and env == "prod":
    client = google.cloud.logging.Client()
    client.setup_logging()

logging.basicConfig(
    level=logging.INFO,
    format="[%(asctime)s] {%(filename)s:%(lineno)d} %(name)s - %(message)s",
)
logger = logging.getLogger(__name__)

if __name__ == "__main__":
    logger.info("Discord Bot Starting")
    bot.run(pycordapi.TOKEN)
