pyproject.toml
[tool.poetry]
name = "discord-gcpai-bot"
version = "0.1.0"
description = ""
authors = [" "]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10, <3.11"

python-dateutil = "^2.8.2"
pandas = "^2.0.0"
db-dtypes = "^1.2.0"
pydantic = "^1.10.4"
requests = "^2.31.0"
discord-webhook  = "1.3.0"
py-cord = "2.4.1"
pillow = "^10.2.0"
emoji = "^2.9.0"

google-cloud-aiplatform = "^1.38.0"
google-cloud-secret-manager = "^2.16.3"
google-cloud-storage = "^2.10.0"
google-cloud-logging = "^3.8.0"
google-cloud-bigquery = "^3.15.0"
gspread = "^6.0.2"

langchain = "0.1.4"
langchain-google-vertexai = "^0.0.3"
langchain-google-genai = "^0.0.6"
langchain-openai= "^0.0.5"
protobuf = "^3.19.3"

python-dotenv = "^0.21.0"

[build-system]
requires = ["poetry-core=1.3.0"]
build-backend = "poetry.core.masonry.api"
